Change how the OpenMRS role runs
Update how the OpenMRS role runs, since during the initial setup there
is some sort of manual intervention that is required. If the
openmrs-runtime.properties file doesn't exist in the server when the
role runs, the role doesn't copy the properties file and create the
trigger in the user_properties table.

Signed-off-by: Jason Rogena <jason@rogena.me>