Run Mybatis migrations as the OpenSRP DB user
Run the Mybatis migrations as the OpenSRP database user once the user
has been given all privileges to the database. This therefore means that
the user doesn't need to be granted access to the schemas (since they're
the ones creating the schemas in the first place).

Also add variables for specifying the OpenSRP tablespace names (instead
of directories). Change was introduced in OpenSRP Server version
v1.2.18-SNAPSHOT.

- https://github.com/OpenSRP/opensrp-server-web/releases/tag/v1.2.18-SNAPSHOT

Related to #88

Signed-off-by: Jason Rogena <jason@rogena.me>