Change the code dir for Mybatis to /opt/opensrp
Install the OpenSRP code in /opt/opensrp instead of /tmp/opensrp in the
application server so that the code can be used to run the Mybatis
migrations.

Do this to cater for prebuilt images since servers that spin up based on
these images will have the /tmp directory empty on-boot.

Signed-off-by: Jason Rogena <jason@rogena.me>