Fix Helm MyBatis migrations issue
In the MyBatis properties file, remove the directive specifying where
the PostgreSQL JDBC driver is to let MyBatis use whatever the OpenSRP
app uses.

It seems the quotes in the .properties file are rendered as being part
of the properties' values. This leads to issues like MyBatis trying to
find a database driver called '"postgresql' instead of 'postgresql'.

Signed-off-by: Jason Rogena <jason@rogena.me>